DROTH-2238 (US 2224) New date fields on traffic signs form have problems with values and validations -Fixed start date can be higher than the end date, and the opposite -Fixed dates not mandatory for traffic sign (142) -Fixed checkbox for additional panels not working -Fixed creating/updating the date property -Fixed datapicker not setting the date to "ei tietoa"

+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/dao/pointasset/OracleTrafficSignDao.scala
@@ -419,8 +419,10 @@ object OracleTrafficSignDao {
         }
       case Date =>
         if (propertyValues.size > 1) throw new IllegalArgumentException("Date property must have exactly one value: " + propertyValues)
-        if (propertyValues.isEmpty) {
+        if (!datePropertyValueDoesNotExist(assetId, propertyId) && propertyValues.head.asInstanceOf[TextPropertyValue].propertyValue.toString.isEmpty) {
           deleteDateProperty(assetId, propertyId).execute
+        } else if (datePropertyValueDoesNotExist(assetId, propertyId) && propertyValues.head.asInstanceOf[TextPropertyValue].propertyValue.toString.isEmpty) {
+          //dont do anything
         } else if (datePropertyValueDoesNotExist(assetId, propertyId)) {
           insertDateProperty(assetId, propertyId, dateFormatter.parseDateTime(propertyValues.head.asInstanceOf[TextPropertyValue].propertyValue.toString)).execute
         } else {
